requirements.txt
# Core dependencies
typer>=0.9.0
rich>=13.0.0
pyyaml>=6.0

# AWS support (optional)
boto3>=1.26.0

# GCP support (optional)
google-cloud-storage>=2.10.0
google-cloud-compute>=1.14.0
google-cloud-logging>=3.5.0
google-cloud-iam>=2.12.0
google-cloud-bigquery>=3.11.0
# google-cloud-sql-connector>=1.0.0  # Not available, using alternative approach
google-cloud-dataproc>=5.4.0

# Azure support (optional)
azure-identity>=1.15.0
azure-mgmt-resource>=23.0.0
azure-mgmt-compute>=30.0.0
azure-mgmt-storage>=21.0.0
azure-mgmt-network>=24.0.0
azure-mgmt-keyvault>=10.2.0
azure-mgmt-security>=5.0.0
azure-mgmt-monitor>=6.0.0
azure-mgmt-sql>=4.0.0
azure-mgmt-datalake-store>=0.5.0
azure-mgmt-databricks>=1.0.0

# AI features (optional)
# openai>=1.0.0
# python-dotenv>=1.0.0

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.5.0
